split out NativeQueryParameterListHelper from NativeQueryImpl
