+++ hibernate-core/src/main/java/org/hibernate/query/sql/internal/NativeQueryImpl.java
@@ -27,12 +27,10 @@ import org.hibernate.LockMode;
 import org.hibernate.Locking;
 import org.hibernate.ScrollMode;
 import org.hibernate.UnknownProfileException;
-import org.hibernate.dialect.Dialect;
 import org.hibernate.engine.query.spi.NativeQueryInterpreter;
 import org.hibernate.engine.spi.SharedSessionContractImplementor;
 import org.hibernate.graph.GraphSemantic;
 import org.hibernate.internal.util.OptionsHelper;
-import org.hibernate.internal.util.MathHelper;
 import org.hibernate.jpa.internal.util.FlushModeTypeHelper;
 import org.hibernate.jpa.spi.NativeQueryArrayTransformer;
 import org.hibernate.jpa.spi.NativeQueryConstructorTransformer;
@@ -80,7 +78,6 @@ import org.hibernate.query.spi.ParameterMetadataImplementor;
 import org.hibernate.query.spi.QueryInterpretationCache;
 import org.hibernate.query.spi.QueryOptions;
 import org.hibernate.query.spi.QueryParameterBindings;
-import org.hibernate.query.spi.QueryParameterImplementor;
 import org.hibernate.query.spi.ScrollableResultsImplementor;
 import org.hibernate.query.spi.SelectQueryPlan;
 import org.hibernate.query.spi.SelectionQueryImplementor;
@@ -91,7 +88,6 @@ import org.hibernate.query.sql.spi.NonSelectInterpretationsKey;
 import org.hibernate.query.sql.spi.ParameterInterpretation;
 import org.hibernate.query.sql.spi.ParameterOccurrence;
 import org.hibernate.query.sql.spi.SelectInterpretationsKey;
-import org.hibernate.sql.ast.spi.ParameterMarkerStrategy;
 import org.hibernate.sql.exec.internal.CallbackImpl;
 import org.hibernate.sql.exec.spi.Callback;
 import org.hibernate.sql.results.graph.Fetchable;
@@ -116,9 +112,7 @@ import java.util.Set;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 
-import static java.lang.Character.isWhitespace;
 import static java.util.Collections.addAll;
-import static org.hibernate.internal.CoreMessageLogger.CORE_LOGGER;
 import static org.hibernate.internal.util.PrimitiveHelper.boxedType;
 import static org.hibernate.internal.util.ReflectHelper.isClass;
 import static org.hibernate.internal.util.StringHelper.unqualify;
@@ -239,7 +233,8 @@ public class NativeQueryImpl<R>
 					if ( specifiedResultType == null && specifiedResultSetMappingName == null ) {
 						// We have neither - SharedSessionContract#createNamedQuery(String) form was called.
 						// Use what's on the selection memento
-						return buildResultSetMapping( getResultSetMappingName( selectionMemento ), false, session );
+						final String registeredName = getResultSetMappingName( selectionMemento );
+						return resolveResultSetMapping( registeredName, false, session.getFactory() );
 					}
 					else if ( specifiedResultSetMappingName != null ) {
 						// One of the forms passing an explicit result-set mapping name was used -
@@ -302,11 +297,13 @@ public class NativeQueryImpl<R>
 		applyMementoOptions( selectionMemento );
 	}
 
-	private static void populateResultSetMappingFromSpecifiedName(String resultSetMappingName, ResultSetMapping resultSetMapping, Consumer<String> querySpaceConsumer, ResultSetMappingResolutionContext context, SharedSessionContractImplementor session) {
-		final var mappingMemento = session.getFactory()
-				.getQueryEngine()
-				.getNamedObjectRepository()
+	private static NamedResultSetMappingMemento resultSetMappingMemento(String resultSetMappingName, SharedSessionContractImplementor session) {
+		return session.getFactory().getQueryEngine().getNamedObjectRepository()
 				.getResultSetMappingMemento( resultSetMappingName );
+	}
+
+	private static void populateResultSetMappingFromSpecifiedName(String resultSetMappingName, ResultSetMapping resultSetMapping, Consumer<String> querySpaceConsumer, ResultSetMappingResolutionContext context, SharedSessionContractImplementor session) {
+		final var mappingMemento = resultSetMappingMemento( resultSetMappingName, session );
 		assert mappingMemento != null;
 		mappingMemento.resolve( resultSetMapping, querySpaceConsumer, context );
 	}
@@ -318,10 +315,8 @@ public class NativeQueryImpl<R>
 			ResultSetMappingResolutionContext context,
 			SharedSessionContractImplementor session) {
 		if ( selectionMemento.getResultMappingName() != null ) {
-			final var resultSetMappingMemento = session.getFactory()
-					.getQueryEngine()
-					.getNamedObjectRepository()
-					.getResultSetMappingMemento( selectionMemento.getResultMappingName() );
+			final var resultSetMappingMemento =
+					resultSetMappingMemento( selectionMemento.getResultMappingName(), session );
 			if ( resultSetMappingMemento != null ) {
 				resultSetMappingMemento.resolve( resultSetMapping, querySpaceConsumer, context );
 				return true;
@@ -1092,7 +1087,7 @@ public class NativeQueryImpl<R>
 
 			@Override
 			public ResultSetMapping getResultSetMapping() {
-				final ResultSetMappingImpl mapping = new ResultSetMappingImpl( "", true );
+				final var mapping = new ResultSetMappingImpl( "", true );
 				mapping.addResultBuilder( new DynamicResultBuilderBasicStandard( 1, longType ) );
 				return mapping;
 			}
@@ -1110,183 +1105,13 @@ public class NativeQueryImpl<R>
 	}
 
 	protected String expandParameterLists(int parameterStartPosition) {
-		if ( parameterOccurrences == null || parameterOccurrences.isEmpty() ) {
-			return sqlString;
-		}
-		// HHH-1123
-		// Some DBs limit number of IN expressions.  For now, warn...
-		final var factory = getSessionFactory();
-		final var dialect = factory.getJdbcServices().getDialect();
-		final boolean paddingEnabled = factory.getSessionFactoryOptions().inClauseParameterPaddingEnabled();
-		final int inExprLimit = dialect.getInExpressionCountLimit();
-		final var parameterMarkerStrategy = factory.getJdbcServices().getParameterMarkerStrategy();
-		final boolean needsMarker = !isStandardRenderer( parameterMarkerStrategy );
-
-		var sql =
-				needsMarker
-						? new StringBuilder( sqlString.length() + parameterOccurrences.size() * 10 )
-								.append( sqlString )
-						: null;
-
-		// Handle parameter lists
-		int offset = 0;
-		int parameterPosition = parameterStartPosition;
-		for ( var occurrence : parameterOccurrences ) {
-			final var queryParameter = occurrence.parameter();
-			final var binding = parameterBindings.getBinding( queryParameter );
-			if ( binding.isMultiValued() ) {
-				final int bindValueCount = binding.getBindValues().size();
-				logTooManyExpressions( inExprLimit, bindValueCount, dialect, queryParameter );
-				final int sourcePosition = occurrence.sourcePosition();
-				if ( sourcePosition >= 0 ) {
-					// check if placeholder is already immediately enclosed in parentheses
-					// (ignoring whitespace)
-					final boolean isEnclosedInParens = isEnclosedInParens( sourcePosition );
-					// short-circuit for performance when only 1 value and the
-					// placeholder is already enclosed in parentheses...
-					if ( bindValueCount != 1 || !isEnclosedInParens ) {
-						if ( sql == null ) {
-							sql = new StringBuilder( sqlString.length() + 20 )
-									.append( sqlString );
-						}
-						final int bindValueMaxCount =
-								determineBindValueMaxCount( paddingEnabled, inExprLimit, bindValueCount );
-						final String expansionListAsString = expandList(
-								bindValueMaxCount,
-								isEnclosedInParens,
-								parameterPosition,
-								parameterMarkerStrategy,
-								needsMarker
-						);
-						final int start = sourcePosition + offset;
-						final int end = start + 1;
-						sql.replace( start, end, expansionListAsString );
-						offset += expansionListAsString.length() - 1;
-						parameterPosition += bindValueMaxCount;
-					}
-					else if ( needsMarker ) {
-						final int start = sourcePosition + offset;
-						final int end = start + 1;
-						final String parameterMarker = parameterMarkerStrategy.createMarker( parameterPosition, null );
-						sql.replace( start, end, parameterMarker );
-						offset += parameterMarker.length() - 1;
-						parameterPosition++;
-					}
-				}
-			}
-			else if ( needsMarker ) {
-				final int sourcePosition = occurrence.sourcePosition();
-				final int start = sourcePosition + offset;
-				final int end = start + 1;
-				final String parameterMarker = parameterMarkerStrategy.createMarker( parameterPosition, null );
-				sql.replace( start, end, parameterMarker );
-				offset += parameterMarker.length() - 1;
-				parameterPosition++;
-			}
-		}
-		return sql == null ? sqlString : sql.toString();
-	}
-
-	private static void logTooManyExpressions(
-			int inExprLimit, int bindValueCount,
-			Dialect dialect, QueryParameterImplementor<?> queryParameter) {
-		if ( inExprLimit > 0 && bindValueCount > inExprLimit ) {
-			CORE_LOGGER.tooManyInExpressions(
-					dialect.getClass().getName(),
-					inExprLimit,
-					queryParameter.getName() == null
-							? queryParameter.getPosition().toString()
-							: queryParameter.getName(),
-					bindValueCount
-			);
-		}
-	}
-
-	private static String expandList(int bindValueMaxCount, boolean isEnclosedInParens, int parameterPosition, ParameterMarkerStrategy parameterMarkerStrategy, boolean needsMarker) {
-		// HHH-8901
-		if ( bindValueMaxCount == 0 ) {
-			return isEnclosedInParens ? "null" : "(null)";
-		}
-		else if ( needsMarker ) {
-			final StringBuilder sb = new StringBuilder( bindValueMaxCount * 4 );
-			if ( !isEnclosedInParens ) {
-				sb.append( '(' );
-			}
-			for ( int i = 0; i < bindValueMaxCount; i++ ) {
-				sb.append( parameterMarkerStrategy.createMarker( parameterPosition + i, null ) );
-				sb.append( ',' );
-			}
-			sb.setLength( sb.length() - 1 );
-			if ( !isEnclosedInParens ) {
-				sb.append( ')' );
-			}
-			return sb.toString();
-		}
-		else {
-			// Shift 1 bit instead of multiplication by 2
-			final char[] chars;
-			if ( isEnclosedInParens ) {
-				chars = new char[(bindValueMaxCount << 1) - 1];
-				chars[0] = '?';
-				for ( int i = 1; i < bindValueMaxCount; i++ ) {
-					final int index = i << 1;
-					chars[index - 1] = ',';
-					chars[index] = '?';
-				}
-			}
-			else {
-				chars = new char[(bindValueMaxCount << 1) + 1];
-				chars[0] = '(';
-				chars[1] = '?';
-				for ( int i = 1; i < bindValueMaxCount; i++ ) {
-					final int index = i << 1;
-					chars[index] = ',';
-					chars[index + 1] = '?';
-				}
-				chars[chars.length - 1] = ')';
-			}
-			return new String( chars );
-		}
-	}
-
-	private boolean isEnclosedInParens(int sourcePosition) {
-		boolean isEnclosedInParens = true;
-		for ( int i = sourcePosition - 1; i >= 0; i-- ) {
-			final char ch = sqlString.charAt( i );
-			if ( !isWhitespace( ch ) ) {
-				isEnclosedInParens = ch == '(';
-				break;
-			}
-		}
-		if ( isEnclosedInParens ) {
-			for ( int i = sourcePosition + 1; i < sqlString.length(); i++ ) {
-				final char ch = sqlString.charAt( i );
-				if ( !isWhitespace( ch ) ) {
-					isEnclosedInParens = ch == ')';
-					break;
-				}
-			}
-		}
-		return isEnclosedInParens;
-	}
-
-	public static int determineBindValueMaxCount(boolean paddingEnabled, int inExprLimit, int bindValueCount) {
-		int bindValueMaxCount = bindValueCount;
-
-		final boolean inClauseParameterPaddingEnabled = paddingEnabled && bindValueCount > 2;
-
-		if ( inClauseParameterPaddingEnabled ) {
-			int bindValuePaddingCount = MathHelper.ceilingPowerOfTwo( bindValueCount );
-
-			if ( inExprLimit > 0 && bindValuePaddingCount > inExprLimit ) {
-				bindValuePaddingCount = inExprLimit;
-			}
-
-			if ( bindValueCount < bindValuePaddingCount ) {
-				bindValueMaxCount = bindValuePaddingCount;
-			}
-		}
-		return bindValueMaxCount;
+		return NativeQueryParameterListHelper.expandParameterLists(
+				sqlString,
+				parameterOccurrences,
+				parameterBindings,
+				getSessionFactory(),
+				parameterStartPosition
+		);
 	}
 
 	private SelectInterpretationsKey selectInterpretationsKey(ResultSetMapping resultSetMapping, int parameterStartPosition) {
@@ -1337,21 +1162,19 @@ public class NativeQueryImpl<R>
 	}
 
 	private NonSelectQueryPlan resolveNonSelectQueryPlan() {
-		NonSelectQueryPlan queryPlan = null;
-
 		final var cacheKey = generateNonSelectInterpretationsKey();
 		if ( cacheKey != null ) {
-			queryPlan = getInterpretationCache().getNonSelectQueryPlan( cacheKey );
-		}
-
-		if ( queryPlan == null ) {
-			final String sqlString = expandParameterLists( 1 );
-			queryPlan = new NativeNonSelectQueryPlanImpl( sqlString, querySpaces, parameterOccurrences );
-			if ( cacheKey != null ) {
-				getInterpretationCache().cacheNonSelectQueryPlan( cacheKey, queryPlan );
+			final var queryPlan = getInterpretationCache().getNonSelectQueryPlan( cacheKey );
+			if ( queryPlan != null ) {
+				return queryPlan;
 			}
 		}
 
+		final String sqlString = expandParameterLists( 1 );
+		final var queryPlan = new NativeNonSelectQueryPlanImpl( sqlString, querySpaces, parameterOccurrences );
+		if ( cacheKey != null ) {
+			getInterpretationCache().cacheNonSelectQueryPlan( cacheKey, queryPlan );
+		}
 		return queryPlan;
 	}
